requirements.txt
google-generativeai>=0.8.0
google-genai>=1.0.0
pypdf2>=3.0.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
click>=8.1.0
scipy>=1.14.0
pydub>=0.25.1